Restructure buidler-docker files and improve it
